refactor(RepositiesTab): use the mediator pattern
Reference implementation of the mediator pattern
on the RepositoriesTab page.

CLOSES DCOS-21638
